@@ -1,16 +1,18 @@
 ﻿using Arcus.EventGrid.Proxy.Api.Extensions;
 using System;
 using System.Linq;
+using System.Text.Json.Serialization;
 using Arcus.EventGrid.Proxy.Api.Validation;
 using Arcus.EventGrid.Publishing;
 using Arcus.EventGrid.Publishing.Interfaces;
+using Arcus.WebApi.Correlation;
+using Arcus.WebApi.Logging;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
+using Microsoft.Extensions.Hosting;
 
 namespace Arcus.EventGrid.Proxy.Api
 {
@@ -29,37 +31,34 @@ namespace Arcus.EventGrid.Proxy.Api
             services.AddMvc()
                 .AddJsonOptions(jsonOptions =>
                 {
-                    jsonOptions.SerializerSettings.Converters.Add(new StringEnumConverter());
-                    jsonOptions.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
+                    jsonOptions.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
+                    jsonOptions.JsonSerializerOptions.IgnoreNullValues = true;
                 })
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                .SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
+
+            services.AddSingleton(BuildEventGridPublisher);
 
             services.UseOpenApiSpecifications();
-            services.AddSingleton(BuildEventGridPublisher);
+            services.AddCorrelation();
+            services.AddHealthChecks();
 
             ValidateConfiguration();
         }
 
-        private void ValidateConfiguration()
-        {
-            var validationOutcomes = RuntimeValidator.Run(Configuration);
-
-            if (validationOutcomes.Any(validationOutcome => validationOutcome.Successful == false))
-            {
-                throw new Exception("Unable to start up due to invalid configuration");
-            }
-        }
-
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
-            app.UseMvc();
+            app.UseMiddleware<ExceptionHandlingMiddleware>();
+            app.UseRouting();
+            app.UseCorrelation();
             app.UseOpenApiDocsWithExplorer();
+
+            app.UseEndpoints(endpoints => endpoints.MapControllers());
         }
 
         private IEventGridPublisher BuildEventGridPublisher(IServiceProvider serviceProvider)
@@ -70,6 +69,16 @@ namespace Arcus.EventGrid.Proxy.Api
             return EventGridPublisherBuilder.ForTopic(topicUri)
                 .UsingAuthenticationKey(authenticationKey)
                 .Build();
+        }
+
+        private void ValidateConfiguration()
+        {
+            var validationOutcomes = RuntimeValidator.Run(Configuration);
+
+            if (validationOutcomes.Any(validationOutcome => validationOutcome.Successful == false))
+            {
+                throw new Exception("Unable to start up due to invalid configuration");
+            }
         }
     }
 }
